fix-separated common components in a folder
