--- src/components/Quiz/SubmitButton/SumbitButton.jsx
@@ -1,9 +1,0 @@
-const SubmitButton = () => {
-    return(
-            <button className="button form-submit-button" type="submit">
-                Sauvegarder ce quiz
-            </button>
-    )
-};
-
-export default SubmitButton;
