+++ src/pages/ExploreByCategory/ExploreByCategory.jsx
@@ -1,0 +1,9 @@
+const ExploreByCategory = () => {
+    return(
+        <>
+            <div>Page Quizzes By Category!</div>
+        </>
+    )
+};
+
+export default ExploreByCategory;
